+++ CHANGELOG.md
@@ -11,6 +11,7 @@ https://github.com/jcorporation/myMPD/
 - Feat: custom placeholder images #980
 - Feat: Fail gracefully if API request fails or is redirected #984
 - Feat: improve queue sort handling #983
+- Feat: use offcanvas instead of popover for contextmenus #974
 - Upd: bootstrap.native 5.0.0-stable
 
 ***